🐛[Fix]: ViewModel에 어울리지 않는 코드 수정
- SetMaximumViewModel에 있던UI 관련 함수 이동
